[refactor] rename TwitterUser to BotPrediction
