+++ app/tests/test_background_task.py
@@ -3,14 +3,14 @@ from datetime import datetime, timedelta
 import pytz  # type: ignore
 
 from app.core.config import settings
-from app.models import TwitterUser
+from app.models import BotPrediction
 from app.services.clean_database import clean_database
 
 MAX_AGE = timedelta(days=settings.RESULT_MAX_AGE)
 
 
 def create_twitter_user(id, timestamp):
-    user_db = TwitterUser(
+    user_db = BotPrediction(
         twitter_id=id,
         tweets_count=0,
         name="user.name",
@@ -53,10 +53,10 @@ def create_fake_twitter_user_collection():
 
 def test_clean_database(client):
     num_new, num_old = create_fake_twitter_user_collection()
-    assert TwitterUser.objects().count() == num_old + num_new
+    assert BotPrediction.objects().count() == num_old + num_new
     clean_database(timedelta(days=settings.RESULT_MAX_AGE))
 
-    users = TwitterUser.objects()
+    users = BotPrediction.objects()
     assert users.count() == num_new
     for user in users:
         assert datetime.now(pytz.UTC) - user.timestamp < MAX_AGE
